[feature] input과 output 추가
